requirements.txt
customtkinter==5.2.2
Pillow==10.3.0
requests==2.31.0
pyinstaller==6.6.0
